Begin TorAppConfig refactor
