+++ app/server/src/main/scala/org/bitcoins/server/BitcoinSServerMain.scala
@@ -114,7 +114,7 @@ class BitcoinSServerMain(override val serverArgParser: ServerArgParser)(implicit
     val feeProvider = getFeeProviderOrElse(
       MempoolSpaceProvider(HourFeeTarget,
                            walletConf.network,
-                           walletConf.torConf.socks5ProxyParams))
+                           Some(torConf.socks5ProxyParams)))
     //get our wallet
     val configuredWalletF = for {
       node <- nodeF
@@ -374,17 +374,20 @@ class BitcoinSServerMain(override val serverArgParser: ServerArgParser)(implicit
         case (None, None) | (None, Some(_)) =>
           default
         case (Some(BitcoinerLive), None) =>
-          BitcoinerLiveFeeRateProvider.fromBlockTarget(6, proxyParams)
+          BitcoinerLiveFeeRateProvider.fromBlockTarget(6, Some(proxyParams))
         case (Some(BitcoinerLive), Some(target)) =>
-          BitcoinerLiveFeeRateProvider.fromBlockTarget(target, proxyParams)
+          BitcoinerLiveFeeRateProvider.fromBlockTarget(target,
+                                                       Some(proxyParams))
         case (Some(BitGo), targetOpt) =>
-          BitGoFeeRateProvider(targetOpt, proxyParams)
+          BitGoFeeRateProvider(targetOpt, Some(proxyParams))
         case (Some(MempoolSpace), None) =>
-          MempoolSpaceProvider(HourFeeTarget, walletConf.network, proxyParams)
+          MempoolSpaceProvider(HourFeeTarget,
+                               walletConf.network,
+                               Some(proxyParams))
         case (Some(MempoolSpace), Some(target)) =>
           MempoolSpaceProvider.fromBlockTarget(target,
                                                walletConf.network,
-                                               proxyParams)
+                                               Some(proxyParams))
         case (Some(Constant), Some(num)) =>
           ConstantFeeRateProvider(SatoshisPerVirtualByte.fromLong(num))
         case (Some(Constant), None) =>
